+++ CHANGES.md
@@ -12,7 +12,8 @@ Known issues:
 
 ## 7.18.0
 
-- TRITON-2510 Add `triton accesskeys` commands to manage access keys.
+- TRITON-2510 Add `triton accesskeys` and `triton rbac accesskey{s}` commands
+  to manage access keys.
 
 ## 7.17.0